# openapi/mindgarden.yaml
openapi: 3.0.3
info:
  title: MindGarden API
  version: '0.1.0'
  description: |
    OpenAPI specification per gli endpoint core di MindGarden: auth, uploads, checkins, gardens, plants, insights, export e settings.
servers:
  - url: https://api.mindgarden.example.com/v1
    description: Production server (example)
  - url: https://staging.api.mindgarden.example.com/v1
    description: Staging
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CheckInCreateText:
      type: object
      properties:
        userId:
          type: string
        timestamp:
          type: string
          format: date-time
        text:
          type: string
        emotionHint:
          type: string
        intensity:
          type: number
        tags:
          type: array
          items:
            type: string
      required: [userId]
    CheckInCreateAudio:
      type: object
      properties:
        userId:
          type: string
        timestamp:
          type: string
          format: date-time
        audioObjectKey:
          type: string
        emotionHint:
          type: string
      required: [userId, audioObjectKey]
    CheckIn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        timestamp:
          type: string
          format: date-time
        text:
          type: string
        sttText:
          type: string
        audioObjectKey:
          type: string
        emotionLabel:
          type: string
        sentimentScore:
          type: number
        intensity:
          type: number
        tags:
          type: array
          items:
            type: string
        embeddingsId:
          type: string
        status:
          type: string
      example:
        id: check-uuid-1234
        userId: user-uuid-1234
        createdAt: '2025-11-19T10:00:00Z'
        timestamp: '2025-11-19T10:00:00Z'
        text: 'Mi sento ansioso per domani'
        sttText: 'Mi sento ansioso per domani'
        emotionLabel: 'ansia'
        sentimentScore: -0.4
        intensity: 0.7
        status: 'complete'
...
paths:
...
  /checkins:
    post:
      summary: Create a check-in (text or audio)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CheckInCreateText'
                - $ref: '#/components/schemas/CheckInCreateAudio'
            examples:
              text-example:
                value:
                  userId: user-uuid-1234
                  text: "Mi sento ansioso per domani"
              audio-example:
                value:
                  userId: user-uuid-1234
                  audioObjectKey: audios/user-uuid-abc123.wav
      responses:
        '201':
          description: Check-in accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkin:
                    $ref: '#/components/schemas/CheckIn'
        '400':
          description: Bad request

    get:
      summary: List check-ins (filterable)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: emotion
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckIn'
  /checkins/{id}:
    get:
      summary: Get check-in detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Check-in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '404':
          description: Not found
...